[BOOKINGSG-9334][GZ] extract time slot component
