+++ src/shared/time-slot/index.ts
@@ -1,0 +1,5 @@
+export type { TimeSlotProps } from "./time-slot";
+export { TimeSlot } from "./time-slot";
+export type { TimeSlotStyleProps } from "./time-slot.styles";
+export { StyledTimeSlot } from "./time-slot.styles";
+export { getCellHeight, getCellWidth } from "./time-slot.styles";
