+++ src/shared/time-slot/types.ts
@@ -1,0 +1,3 @@
+export type SlotStyle = "default" | "stripes";
+
+export type Direction = "right" | "left";
