+++ src/time-slot-bar/time-slot-bar.styles.tsx
@@ -3,34 +3,17 @@ import { ChevronRightIcon } from "@lifesg/react-icons/chevron-right";
 import styled, { css } from "styled-components";
 
 import { ClickableIcon } from "../shared/clickable-icon";
+import { getCellHeight, getCellWidth } from "../shared/time-slot";
+import type { Direction } from "../shared/time-slot/types";
 import { Typography } from "../typography";
 import { V3_Border, V3_Colour, V3_Shadow } from "../v3_theme";
-import type { V3_ThemeStyleProps } from "../v3_theme/types";
-import type { Direction, SlotStyle, TimeSlotBarVariant } from "./types";
+import type { TimeSlotBarVariant } from "./types";
 
 const MAX_LINE_HEIGHT = 1.25; // NOTE in rem
 
 // =============================================================================
 // STYLING HELPERS
 // =============================================================================
-
-// Function to get the width of a cell in px
-export const getCellWidth = (variant: TimeSlotBarVariant) => {
-    if (variant === "minified") {
-        return 12;
-    } else {
-        return 40;
-    }
-};
-
-// Function to get the height of a cell in px
-export const getCellHeight = (variant: TimeSlotBarVariant) => {
-    if (variant === "minified") {
-        return 12;
-    } else {
-        return 40;
-    }
-};
 
 // =============================================================================
 // STYLE INTERFACE, transient props are denoted with $
@@ -40,19 +23,6 @@ export const getCellHeight = (variant: TimeSlotBarVariant) => {
 interface ArrowStyleProps {
     $direction?: Direction;
     $variant: TimeSlotBarVariant;
-}
-
-export interface TimeSlotStyleProps {
-    $type?: "default" | "vertical";
-    $variant: TimeSlotBarVariant;
-    $width?: number;
-    $left?: number;
-    $styleType: SlotStyle;
-    $bgColor: string | ((props: V3_ThemeStyleProps) => string);
-    $bgColor2?: string | ((props: V3_ThemeStyleProps) => string);
-    $hoverBgColor?: string | ((props: V3_ThemeStyleProps) => string);
-    $hoverBgColor2?: string | ((props: V3_ThemeStyleProps) => string);
-    $clickable?: boolean;
 }
 
 interface CellTextStyleProps {
@@ -169,75 +139,6 @@ export const TimeLabel = styled(Typography.BodyXS)`
     position: absolute;
     bottom: 10%;
     left: 10%;
-`;
-
-export const TimeSlot = styled.div<TimeSlotStyleProps>`
-    ${(props) => {
-        if (props.$type === "vertical") {
-            return css`
-                display: flex;
-                flex-grow: 1;
-                align-items: center;
-                justify-content: center;
-                width: 100%;
-                margin: 1px 0px;
-            `;
-        } else {
-            return css`
-                position: absolute;
-                height: ${getCellHeight(props.$variant)}px;
-                width: ${props.$width}px;
-                left: ${props.$left}px;
-            `;
-        }
-    }}
-    background-color: ${({ $bgColor }) => $bgColor};
-    cursor: ${({ $clickable }) => ($clickable ? "pointer" : "default")};
-    ${({ $hoverBgColor, $clickable }) =>
-        $hoverBgColor &&
-        $clickable &&
-        css`
-            &:hover {
-                background-color: ${$hoverBgColor};
-            }
-        `}
-
-    ${(props) =>
-        props.$styleType === "stripes" &&
-        css`
-            background: repeating-linear-gradient(
-                135deg,
-                ${props.$bgColor2 || V3_Colour["bg-strongest"]} 0px,
-                ${props.$bgColor2 || V3_Colour["bg-strongest"]} 10px,
-                ${props.$bgColor || V3_Colour["bg-stronger"]} 10px,
-                ${props.$bgColor || V3_Colour["bg-stronger"]} 20px
-            );
-            ${(props.$hoverBgColor || props.$hoverBgColor2) &&
-            props.$clickable &&
-            css`
-                &:hover {
-                    background: repeating-linear-gradient(
-                        135deg,
-                        ${props.$hoverBgColor2 ||
-                            props.$bgColor2 ||
-                            V3_Colour["bg-strongest"]}
-                            0px,
-                        ${props.$hoverBgColor2 ||
-                            props.$bgColor2 ||
-                            V3_Colour["bg-strongest"]}
-                            10px,
-                        ${props.$hoverBgColor ||
-                            props.$bgColor ||
-                            V3_Colour["bg-stronger"]}
-                            10px,
-                        ${props.$hoverBgColor ||
-                            props.$bgColor ||
-                            V3_Colour["bg-stronger"]}
-                            20px
-                    );
-                }
-            `}
-        `}
 `;
 
 export const TimeSlotBorder = styled.div<{ $variant: TimeSlotBarVariant }>`
