+++ src/time-slot-bar/types.ts
@@ -1,8 +1,5 @@
+import type { SlotStyle } from "../shared/time-slot/types";
 import type { V3_ThemeStyleProps } from "../v3_theme/types";
-
-export type SlotStyle = "default" | "stripes";
-
-export type Direction = "right" | "left";
 
 export type TimeSlotBarVariant = "minified" | "default";
